+++ src/hardware/hw_main.h
@@ -33,8 +33,8 @@ void HWR_Shutdown(void);
 
 void HWR_clearAutomap(void);
 void HWR_drawAMline(const fline_t *fl, INT32 color);
-void HWR_FadeScreenMenuBack(ULONG color, INT32 height);
-void HWR_DrawConsoleBack(ULONG color, INT32 height);
+void HWR_FadeScreenMenuBack(UINT32 color, INT32 height);
+void HWR_DrawConsoleBack(UINT32 color, INT32 height);
 void HWR_RenderPlayerView(INT32 viewnumber, player_t *player);
 void HWR_DrawViewBorder(INT32 clearlines);
 void HWR_DrawFlatFill(INT32 x, INT32 y, INT32 w, INT32 h, lumpnum_t flatlumpnum);
